Parte 2 Examen final

+++ Core/Inc/room_control.h
@@ -47,7 +47,7 @@ typedef struct {
 void room_control_init(room_control_t *room);
 void room_control_update(room_control_t *room);
 void room_control_process_key(room_control_t *room, char key);
-void room_control_set_temperature(room_control_t *room, float temperature);
+void room_control_set_temperature(room_control_t *room, int temperature);
 void room_control_force_fan_level(room_control_t *room, fan_level_t level);
 void room_control_change_password(room_control_t *room, const char *new_password);
